+++ src/config.cpp
@@ -222,6 +222,7 @@ comment_config::comment_config() : cmd_char_('\\')
 
     STANDARDESE_DETAIL_SET(exclude)
     STANDARDESE_DETAIL_SET(unique_name)
+    STANDARDESE_DETAIL_SET(synopsis)
     STANDARDESE_DETAIL_SET(group)
     STANDARDESE_DETAIL_SET(entity)
     STANDARDESE_DETAIL_SET(file)